finished location parsing
