+++ tests/true_nwbinspector_log_file_testing.txt
@@ -3,22 +3,24 @@ NWBFile: xxxxxx.nwb
 
 CRITICAL IMPORTANCE
 -------------------
-1.0.1   TimeSeries 'test_time_series_4' located in 'TODO'
+1.0.1   TimeSeries 'test_time_series_3' located in '/nwbfile/acquisition/'
         check_timestamps_match_first_dimension: The length of the first dimension of data does not match the length of timestamps.
-1.0.2   TimeSeries 'test_time_series_4' located in 'TODO'
+1.0.2   TimeSeries 'test_time_series_3' located in '/nwbfile/acquisition/'
+        check_data_orientation: Data may be in the wrong orientation. Time should be in the first dimension, and is usually the longest dimension. Here, another dimension is longer. 
+1.0.3   SpatialSeries 'my_spatial_series' located in '/nwbfile/processing/behavior/Position/'
         check_data_orientation: Data may be in the wrong orientation. Time should be in the first dimension, and is usually the longest dimension. Here, another dimension is longer. 
 
 BEST PRACTICE VIOLATION
 -----------------------
-1.1.1   TimeSeries 'test_time_series_1' located in 'TODO'
+1.1.1   TimeSeries 'test_time_series_1' located in '/nwbfile/acquisition/'
         check_dataset_compression: Consider enabling compression when writing a large dataset.
-1.1.2   TimeSeries 'test_time_series_2' located in 'TODO'
+1.1.2   TimeSeries 'test_time_series_2' located in '/nwbfile/acquisition/'
         check_regular_timestamps: TimeSeries appears to have a constant sampling rate. Consider specifying starting_time=1.2 and rate=2.0 instead of timestamps.
-1.1.3   TimeSeries 'test_time_series_4' located in 'TODO'
+1.1.3   TimeSeries 'test_time_series_3' located in '/nwbfile/acquisition/'
         check_regular_timestamps: TimeSeries appears to have a constant sampling rate. Consider specifying starting_time=1.0 and rate=1.0 instead of timestamps.
-1.1.4   TimeSeries 'test_time_series_2' located in 'TODO'
+1.1.4   TimeSeries 'test_time_series_2' located in '/nwbfile/acquisition/'
         check_dataset_compression: Consider enabling compression when writing a large dataset.
-1.1.5   TimeSeries 'test_time_series_3' located in 'TODO'
+1.1.5   TimeSeries 'test_time_series_3' located in '/nwbfile/acquisition/'
         check_dataset_compression: Consider enabling compression when writing a large dataset.
-1.1.6   TimeSeries 'test_time_series_4' located in 'TODO'
+1.1.6   SpatialSeries 'my_spatial_series' located in '/nwbfile/processing/behavior/Position/'
         check_dataset_compression: Consider enabling compression when writing a large dataset.
